LORDstar: Asychronous online mFDR control based on recent discovery

Implements LORD algorithms for asynchronous online testing, as presented by Zrnic et al. (2018).

LORDstar(pval, alpha = 0.05, version, gammai, w0, decision.times, lags,
  batch.sizes)

Arguments

pval
A vector of p-values.

alpha
Overall significance level of the procedure, the default is 0.05.

version
Takes values 'async', 'dep' or 'batch'. This specifies the version of LORDstar to use.

gammai
Optional vector of \(\gamma_i\). A default is provided as proposed by Javanmard and Montanari (2018), equation 31.

w0
Initial `wealth' of the procedure, defaults to \(\alpha/10\).

decision.times
A vector of decision times for the hypothesis tests, this is required for version='async'.

lags
A vector of lags or the hypothesis tests, this is required for version='dep'.

batch.sizes
A vector of batch sizes, this is required for version='batch'.

Value

d.out

A dataframe with the original p-values pval, the adjusted testing levels \(\alpha_i\) and the indicator function of discoveries R. Hypothesis \(i\) is rejected if the \(i\)-th p-value is less than or equal to \(\alpha_i\), in which case R[i] = 1 (otherwise R[i] = 0).

Details

The function takes as its input a vector of p-values, as well as a vector describing the conflict sets for the hypotheses. This takes the form of a vector of decision times, lags or batch sizes (see below).

Zrnic et al. (2018) present explicit three versions of LORDstar:

1) version='async' is for an asynchoronous testing process, consisting of tests that start and finish at (potentially) random times. The discretised finish times of the test correspond to the decision times. These decision times are given as the input decision.times for this version of the LORDstar algorithm.

2) version='dep' is for online testing under local dependence of the p-values. More precisely, for any \(t>0\) we allow the p-value \(p_t\) to have arbitrary dependence on the previous \(L_t\) p-values. The fixed sequence \(L_t\) is referred to as `lags', and is given as the input lags for this version of the LORDstar algorithm.

3) version='batch' is for controlling the mFDR in mini-batch testing, where a mini-batch represents a grouping of tests run asynchronously which result in dependent p-values. Once a mini-batch of tests is fully completed, a new one can start, testing hypotheses independent of the previous batch. The batch sizes are given as the input batch.sizes for this version of the LORDstar algorithm.

Given an overall significance level \(\alpha\), LORDstar depends on \(w_0\) (where \(0 \le w_0 \le \alpha\)), which represents the intial `wealth' of the procedure. The algorithms also require a sequence of non-negative non-increasing numbers \(\gamma_i\) that sum to 1.

Note that these LORDstar algorithms control the modified FDR (mFDR). The `async' version also controls the usual FDR if the p-values are assumed to be independent.

Further details of the LORDstar algorithms can be found in Zrnic et al. (2018).

References

Javanmard, A. and Montanari, A. (2018) Online Rules for Control of False Discovery Rate and False Discovery Exceedance. Annals of Statistics, 46(2):526-554.

Zrnic, T., Ramdas, A. and Jordan, M.I. (2018). Asynchronous Online Testing of Multiple Hypotheses. arXiv preprint, https://arxiv.org/abs/1812.05068.

See also

Examples

pval = c(2.90e-08, 0.06743, 3.51e-04, 0.00174, 0.04723,
        3.60e-05, 0.79149, 0.27201, 0.28295, 7.59e-06,
        0.69274, 0.30443, 0.00136, 0.82342, 0.54757)

LORDstar(pval, version='async', decision.times=seq_len(15)) # Synchronous

#>          pval       alphai R
#> 1  2.9000e-08 0.0002675839 1
#> 2  6.7430e-02 0.0024664457 0
#> 3  3.5100e-04 0.0005732818 1
#> 4  1.7400e-03 0.0031631190 1
#> 5  4.7230e-02 0.0036636555 0
#> 6  3.6000e-05 0.0014222639 1
#> 7  7.9149e-01 0.0038823065 0
#> 8  2.7201e-01 0.0016066352 0
#> 9  2.8295e-01 0.0013823033 0
#> 10 7.5900e-06 0.0011925815 1
#> 11 6.9274e-01 0.0037218248 0
#> 12 3.0443e-01 0.0015130314 0
#> 13 1.3600e-03 0.0013346425 0
#> 14 8.2342e-01 0.0011757579 0
#> 15 5.4757e-01 0.0010500975 0

LORDstar(pval, version='async', decision.times=seq_len(15)+1) # Asynchronous

#>          pval       alphai R
#> 1  2.9000e-08 2.675839e-04 1
#> 2  6.7430e-02 5.819103e-05 0
#> 3  3.5100e-04 5.732818e-04 1
#> 4  1.7400e-03 4.872805e-04 0
#> 5  4.7230e-02 9.878169e-04 0
#> 6  3.6000e-05 8.403536e-04 1
#> 7  7.9149e-01 7.108430e-04 0
#> 8  2.7201e-01 1.194455e-03 0
#> 9  2.8295e-01 1.032860e-03 0
#> 10 7.5900e-06 8.902865e-04 1
#> 11 6.9274e-01 7.800141e-04 0
#> 12 3.0443e-01 1.275770e-03 0
#> 13 1.3600e-03 1.120595e-03 0
#> 14 8.2342e-01 9.808453e-04 0
#> 15 5.4757e-01 8.712179e-04 0

LORDstar(pval, version='dep', lags=rep(0,15)) # Synchronous

#>          pval lag       alphai R
#> 1  2.9000e-08   0 0.0002675839 1
#> 2  6.7430e-02   0 0.0024664457 0
#> 3  3.5100e-04   0 0.0005732818 1
#> 4  1.7400e-03   0 0.0031631190 1
#> 5  4.7230e-02   0 0.0036636555 0
#> 6  3.6000e-05   0 0.0014222639 1
#> 7  7.9149e-01   0 0.0038823065 0
#> 8  2.7201e-01   0 0.0016066352 0
#> 9  2.8295e-01   0 0.0013823033 0
#> 10 7.5900e-06   0 0.0011925815 1
#> 11 6.9274e-01   0 0.0037218248 0
#> 12 3.0443e-01   0 0.0015130314 0
#> 13 1.3600e-03   0 0.0013346425 0
#> 14 8.2342e-01   0 0.0011757579 0
#> 15 5.4757e-01   0 0.0010500975 0

LORDstar(pval, version='dep', lags=rep(1,15)) # Locally dependent

#>          pval lag       alphai R
#> 1  2.9000e-08   1 2.675839e-04 1
#> 2  6.7430e-02   1 5.819103e-05 0
#> 3  3.5100e-04   1 5.732818e-04 1
#> 4  1.7400e-03   1 4.872805e-04 0
#> 5  4.7230e-02   1 9.878169e-04 0
#> 6  3.6000e-05   1 8.403536e-04 1
#> 7  7.9149e-01   1 7.108430e-04 0
#> 8  2.7201e-01   1 1.194455e-03 0
#> 9  2.8295e-01   1 1.032860e-03 0
#> 10 7.5900e-06   1 8.902865e-04 1
#> 11 6.9274e-01   1 7.800141e-04 0
#> 12 3.0443e-01   1 1.275770e-03 0
#> 13 1.3600e-03   1 1.120595e-03 0
#> 14 8.2342e-01   1 9.808453e-04 0
#> 15 5.4757e-01   1 8.712179e-04 0

LORDstar(pval, version='batch', batch.sizes=rep(1,15)) # Synchronous

#>          pval batch       alphai R
#> 1  2.9000e-08     1 0.0002675839 1
#> 2  6.7430e-02     2 0.0024664457 0
#> 3  3.5100e-04     3 0.0005732818 1
#> 4  1.7400e-03     4 0.0031631190 1
#> 5  4.7230e-02     5 0.0036636555 0
#> 6  3.6000e-05     6 0.0014222639 1
#> 7  7.9149e-01     7 0.0038823065 0
#> 8  2.7201e-01     8 0.0016066352 0
#> 9  2.8295e-01     9 0.0013823033 0
#> 10 7.5900e-06    10 0.0011925815 1
#> 11 6.9274e-01    11 0.0037218248 0
#> 12 3.0443e-01    12 0.0015130314 0
#> 13 1.3600e-03    13 0.0013346425 0
#> 14 8.2342e-01    14 0.0011757579 0
#> 15 5.4757e-01    15 0.0010500975 0

LORDstar(pval, version='batch', batch.sizes=c(4,6,5)) # Batched

#>          pval batch       alphai R
#> 1  2.9000e-08     1 2.675839e-04 1
#> 2  6.7430e-02     1 5.819103e-05 0
#> 3  3.5100e-04     1 4.956249e-05 0
#> 4  1.7400e-03     1 4.121803e-05 0
#> 5  4.7230e-02     2 2.443199e-03 0
#> 6  3.6000e-05     2 5.539488e-04 1
#> 7  7.9149e-01     2 4.726597e-04 0
#> 8  2.7201e-01     2 3.946884e-04 0
#> 9  2.8295e-01     2 3.359039e-04 0
#> 10 7.5900e-06     2 2.915568e-04 1
#> 11 6.9274e-01     3 5.608940e-03 0
#> 12 3.0443e-01     3 1.393881e-03 0
#> 13 1.3600e-03     3 1.199247e-03 0
#> 14 8.2342e-01     3 1.014118e-03 0
#> 15 5.4757e-01     3 8.733223e-04 0