Deprecated functions in package ‘onlineFDR’

These functions are provided for compatibility with older versions of ‘onlineFDR’ only, and will be defunct at the next release.

Details

The following functions are deprecated and will be made defunct; use the replacement indicated below:

LORDdep: LORD with version='dep'

bonfInfinite: Alpha_spending